tracks and demo
Both tracks and uncorrelated plots and historic tracks and plots now works decently, after my own mind. Don't know how this works on the real system in detail, so it is only guesswork in the pursuit of giving a good show. The correct number of historic tracks is now shown. New demo recording done (2 AJS Viggen and 1 C130 landing).